Using transducer capabilities of functions
Now we simplify our implemnentation by redefining process-culsters to
use the transducer facility of mapcat, filter and map.

The partial function is removed from each of these functions, so each
one returns a transducer.  In fact the whole process-clusters becomes a
transducer.

The process-cluster can now be used with sequence functions such as
into, sequence and reduce, simplifying each of these calls.